User POJO moved to "User" folder
