+++ src/main/java/ru/pojo/User/UserLogin.java
@@ -1,19 +1,16 @@
-package ru.pojo;
+package ru.pojo.User;
 
-public class UserRegister {
+public class UserLogin {
     private String email;
     private String password;
-    private String name;
 
-    public UserRegister(){
+    public UserLogin(){
 
     }
 
-    public UserRegister(String email, String password, String name){
+    public UserLogin(String email, String password){
         this.email = email;
         this.password = password;
-        this.name = name;
-
     }
 
     public String getEmail() {
@@ -23,7 +20,6 @@ public class UserRegister {
     public void setEmail(String email) {
         this.email = email;
     }
-
     public String getPassword() {
         return password;
     }
@@ -31,12 +27,5 @@ public class UserRegister {
     public void setPassword(String password) {
         this.password = password;
     }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
 }
+
