+++ src/main/java/ru/pojo/User/UserRegister.java
@@ -1,16 +1,19 @@
-package ru.pojo;
+package ru.pojo.User;
 
-public class UserLogin {
+public class UserRegister {
     private String email;
     private String password;
+    private String name;
 
-    public UserLogin(){
+    public UserRegister(){
 
     }
 
-    public UserLogin(String email, String password){
+    public UserRegister(String email, String password, String name){
         this.email = email;
         this.password = password;
+        this.name = name;
+
     }
 
     public String getEmail() {
@@ -20,6 +23,7 @@ public class UserLogin {
     public void setEmail(String email) {
         this.email = email;
     }
+
     public String getPassword() {
         return password;
     }
@@ -27,5 +31,12 @@ public class UserLogin {
     public void setPassword(String password) {
         this.password = password;
     }
-}
 
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
